server/requirements.txt
asgiref==3.11.0
channels==4.3.2
Django==6.0.1
django-cors-headers==4.9.0
django-rest-framework==0.1.0
django-shortuuidfield==0.1.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.0
PyJWT==2.10.1
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.5
